Fix PWM bug
I was using the wrong operator while assigning to the data direction
register for the steering and throttle pins. I should've used the
OR-assignment (|=) instead of the regular assignment (=). When I
did the regular assignment of the steering DDR pin after setting the
throttle DDR pin, I effectively overwrote the throttle DDR pin. Thus,
the throttle PWM waveform was not being created. This is now fixed!